adds godoc for request qualifier methods
